Arg, pycharm, why
I had some git trouble in pycharm editing pygge when it was open as a content root for another project. Thus, this is two commits bunched up. The key thing is that I simplified a little bit the constructors. It's still not perfect, but at least we get rid of some unnecessarily complicated attribute-dict games.